+++ example/QueueDoctrine/config/module.config.php
@@ -7,7 +7,6 @@ return [
         'factories' => [
             Manager\EntityManager::class => \Zend\ServiceManager\AbstractFactory\ConfigAbstractFactory::class,
             \Bupy7\Queue\Service\QueueService::class => \QueueDoctrine\Service\QueueServiceFactory::class,
-            \Bupy7\Queue\Service\TaskService::class => \QueueDoctrine\Service\TaskServiceFactory::class,
         ],
     ],
     \Zend\ServiceManager\AbstractFactory\ConfigAbstractFactory::class => [