+++ example/QueueDoctrine/src/Module.php
@@ -3,8 +3,12 @@
 namespace QueueDoctrine;
 
 use Zend\ModuleManager\Feature\ConfigProviderInterface;
+use Zend\ModuleManager\Feature\BootstrapListenerInterface;
+use Zend\EventManager\EventInterface;
+use QueueDoctrine\Type\ParamsType;
+use Doctrine\DBAL\Types\Type;
 
-class Module implements ConfigProviderInterface
+class Module implements ConfigProviderInterface, BootstrapListenerInterface
 {
     public function getConfig(): array
     {
@@ -12,5 +16,12 @@ class Module implements ConfigProviderInterface
             require __DIR__ . '/../config/module.config.php',
             require __DIR__ . '/../config/doctrine.config.php'
         );
+    }
+
+    public function onBootstrap(EventInterface $e)
+    {
+        if (!Type::hasType(ParamsType::PARAMS)) {
+            Type::addType(ParamsType::PARAMS, ParamsType::class);
+        }
     }
 }