+++ test/Functional/_file/config/module.config.php
@@ -7,12 +7,12 @@ return [
         'factories' => [
             Assert\Task\ErrorHelloTask::class => \Zend\ServiceManager\Factory\InvokableFactory::class,
             Assert\Task\SuccessHelloTask::class => \Zend\ServiceManager\Factory\InvokableFactory::class,
+            Assert\Task\ExceptionHelloTask::class => \Zend\ServiceManager\Factory\InvokableFactory::class,
         ],
     ],
     'service_manager' => [
         'factories' => [
             'MemoryQueueService' => Assert\Service\MemoryQueueServiceFactory::class,
-            'MemoryTaskService' => Assert\Service\MemoryTaskServiceFactory::class,
 
             Assert\Repository\MemoryTaskRepository::class => \Zend\ServiceManager\Factory\InvokableFactory::class,
             Assert\Repository\MemoryExceptionTaskRepository::class