+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 zf-queue
 ========
 
+1.0.0-alpha.2
+-------------
+
+- Enh: Drop `Bupy7\Queue\Service\TaskService` and
+moved `Bupy7\Queue\Service\TaskService::add()` method
+to `Bupy7\Queue\Service\QueueService::add()`.
+
 1.0.0-alpha
 -----------